Move to using a src folder
