--- db.py
@@ -1,3 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy # type: ignore
-
-db = SQLAlchemy()
