--- models/__init__.py
@@ -1,1 +1,0 @@
-from models.rank import RankModel # type: ignore
