+++ pyproject.toml
@@ -5,9 +5,22 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
+    "build>=1.3.0",
     "flask>=3.1.2",
     "flask-migrate>=4.1.0",
     "load-dotenv>=0.1.0",
     "psycopg2-binary>=2.9.10",
     "sqlalchemy>=2.0.43",
 ]
+
+[build-system]
+requires = ["setuptools"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+
