+++ src/api/routes.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+bp = Blueprint("main", __name__)
+
+@bp.route("/")
+def hello():
+    return "Hello from Flask!"
